Rename folder 'alu' -> 'arithmetic'.
