+++ evm/src/arithmetic/decode.rs
